requirements.txt
# Core dependencies
langgraph>=0.0.40
google-generativeai>=0.3.0
psycopg2-binary>=2.9.7
requests>=2.31.0

# Development and testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0

# Optional dependencies for enhanced functionality
fastapi>=0.103.0
uvicorn>=0.23.0
pydantic>=2.0.0
jinja2>=3.1.0

# Documentation
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0

grandalf
